Started on login
Kinda works as a foundation. Whatever the user types into the user box
on the login page will result in the browser storing a cookie on what
user is logged in as an element, but I'm still figureing out how to use
this bit of data for profiles and other usefull stuff.

Good progress though! :-1: